Add cast to remove compiler warning (causing build break)
Reviewed-on: http://review.couchbase.org/30399
Tested-by: Trond Norbye <trond.norbye@gmail.com>

+++ src/checkpoint.cc
@@ -191,7 +191,7 @@ size_t Checkpoint::mergePrevCheckpoint(Checkpoint *pPrevCheckpoint) {
             // Skip the first two meta items
             ++pos; ++pos;
             toWrite.insert(pos, *rit);
-            index_entry entry = {--pos, pPrevCheckpoint->getMutationIdForKey(key)};
+            index_entry entry = {--pos, static_cast<int64_t>(pPrevCheckpoint->getMutationIdForKey(key))};
             keyIndex[key] = entry;
             newEntryMemOverhead += key.size() + sizeof(index_entry);
             ++numItems;
